+++ models/linkedin_ads__creative_report.sql
@@ -38,7 +38,7 @@ final as (
 
     select 
         report.date_day,
-        creative.creative_id,
+        report.creative_id,
         campaign.campaign_id,
         campaign.campaign_name,
         campaign_group.campaign_group_id,
@@ -69,7 +69,7 @@ final as (
     left join account 
         on campaign.account_id = account.account_id
 
-    {{ dbt_utils.group_by(n=18) }}
+    {{ dbt_utils.group_by(n=16) }}
 
 )