+++ models/linkedin_ads__url_report.sql
@@ -75,7 +75,7 @@ final as (
         where creative.click_uri is not null
     {% endif %}
 
-    {{ dbt_utils.group_by(n=19) }}
+    {{ dbt_utils.group_by(n=18) }}
 
 )